nest *-layer.js files in layer dir move helper functions out of neighborhoods-layer to util.js
